requirements.txt
fastapi
motor
python-dotenv
passlib[bcrypt]
pydantic
python-jose[cryptography]
uvicorn